Fix broken module when the local time changes during daily forecasts.
This patch aims to fix issue #47.

It replaces the whole Date processing using moment for daily forecasts.
Processing is also simplified to avoid looping around the forecast data
multiple time.

The selected time of the target day used for the daily forecast is no
longer set to 12:00 but selected in the middle of forecast data.

Signed-off-by: Romain Bezut <morian@xdec.net>